+++ apollo-router/src/plugins/response_cache/mod.rs
@@ -4,8 +4,8 @@ pub(crate) mod invalidation;
 pub(crate) mod invalidation_endpoint;
 pub(crate) mod metrics;
 pub(crate) mod plugin;
-pub(crate) mod postgres;
 pub(crate) mod serde_blake3;
+mod storage;
 #[cfg(all(
     test,
     any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux"))
