+++ apollo-router/src/plugins/response_cache/storage/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod postgres;
